fix block writer flushing
Flushing block writer writes any remaining bits from the internal buffer
to the underlying writer. If there is no bits for the full byte they are
also flushed resulting in partial byte. It is safe to use this only when
we are on the byte boundary or when we are at the end of the final block
and we had to flush all bits.

Positioning on the byte boundary is possible with stored block. Which
is (after block type) padded to the byte boundary. So if we want to
flush any internal buffers we had to write 0 length stored block and
then flush internal bit buffer.

Fixes: #9