update benchmark results
Previous version was highly inlined and I was comparing in ReleaseSafe
mode. Those inlining turned ReleaseSafe into ReleaseFast for this
library leaving std in ReleaseSafe. Which is not fer comparison. This
benchmark if after removing most of the inlines, because they don't make
any difference in ReleaseFast and comparing now both in ReleaseFast.